+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `DatadogLogger` will no longer leak its reference to its native Logger.
 * Fix debug output from native `DatadogSdk` on iOS.
+* Fix `LogLevel` being private. See [#518]
 
 ## 2.0.0
 
@@ -209,3 +210,4 @@ Release 2.0 introduces breaking changes. Follow the [Migration Guide](MIGRATING.
 [#416]: https://github.com/DataDog/dd-sdk-flutter/issues/416
 [#462]: https://github.com/DataDog/dd-sdk-flutter/issues/462
 [#472]: https://github.com/DataDog/dd-sdk-flutter/issues/472
+[#518]: https://github.com/DataDog/dd-sdk-flutter/issues/518